:white_check_mark: InformaciónPersonalRest + Mapper + correcciones

+++ workspace/EC1ConceptualRelacional/src/main/java/com/cabos/controller/InformacionPersonalRestController.java
@@ -2,7 +2,6 @@ package com.cabos.controller;
 
 import java.util.Collection;
 
-
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -15,6 +14,8 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
+import com.cabos.mapper.InfoPersonalMapper;
+import com.cabos.mapper.MapperUtil;
 import com.cabos.model.InformacionPersonal;
 import com.cabos.services.InformacionPersonalService;
 
@@ -24,34 +25,38 @@ public class InformacionPersonalRestController {
 
 	@Autowired
 	private InformacionPersonalService service;
+	
+	@GetMapping("/listar")
+	public ResponseEntity<?> listar(){
+		Collection<InformacionPersonal> infosPersonales = service.findAll();
 
+		if(infosPersonales.isEmpty())
+			return new ResponseEntity<>("No hay informaciones personales.",HttpStatus.NO_CONTENT);
+		
+		Collection<InfoPersonalMapper> infosPersonalesMapper = MapperUtil.convertCollInfoP(infosPersonales);
+		
+		return new ResponseEntity<>(infosPersonalesMapper,HttpStatus.OK);
+	}
 	
 	@PostMapping("/agregar")
 	public ResponseEntity<?> agregar(@RequestBody InformacionPersonal informacionPersonal){
 		service.insert(informacionPersonal);
-		return new ResponseEntity<>(HttpStatus.CREATED);
+		return new ResponseEntity<>(
+				"¡Información personal ID: "+informacionPersonal.getIdInfoPersonal()+", "+
+				"\""+informacionPersonal.getNombres()+" "+informacionPersonal.getApellidos()+"\"!",
+				HttpStatus.CREATED);
 	}
-	
-
-	@GetMapping("/listar")
-	public ResponseEntity<?> listar(){
-		Collection<InformacionPersonal> infoPersonales = service.findAll();
-		
-		if(infoPersonales.isEmpty())
-			return new ResponseEntity<>(HttpStatus.NO_CONTENT);
-		
-		return new ResponseEntity<>(infoPersonales,HttpStatus.OK);
-	}
-	
 	
 	@GetMapping("/buscar/{idInfoPersonal}")
 	public ResponseEntity<?> buscar(@PathVariable Integer idInfoPersonal){
 		InformacionPersonal infoPersonalDb = service.findById(idInfoPersonal);
 		
 		if(infoPersonalDb!=null)
-			return new ResponseEntity<>(infoPersonalDb,HttpStatus.OK);
+			return new ResponseEntity<>(MapperUtil.convertInfoPersonal(infoPersonalDb) ,HttpStatus.OK);
 		
-		return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+		return new ResponseEntity<>(
+				"¡No existe información personal con el ID: "+idInfoPersonal+"!",
+				HttpStatus.NOT_FOUND);
 	}
 	
 	@PutMapping("/editar/{idInfoPersonal}")
@@ -66,9 +71,11 @@ public class InformacionPersonalRestController {
 			infoPersonalDb.setCorreo(informacionPersonal.getCorreo());
 			infoPersonalDb.setfNacimiento(informacionPersonal.getfNacimiento());
 			infoPersonalDb.setNombres(informacionPersonal.getNombres());
-			service.insert(infoPersonalDb);			
+			service.update(infoPersonalDb);			
 			
-			return new ResponseEntity<>(HttpStatus.OK);
+			return new ResponseEntity<>(
+					"¡No existe información personal con el ID: "+idInfoPersonal+"!",
+					HttpStatus.OK);
 		}
 		
 		return new ResponseEntity<>("El código "+idInfoPersonal+"no existe",HttpStatus.NOT_FOUND);
@@ -76,7 +83,16 @@ public class InformacionPersonalRestController {
 	
 	@DeleteMapping("/eliminar/{idInfoPersonal}")
 	public ResponseEntity<?> eliminar(@PathVariable Integer idInfoPersonal){
-		service.delete(idInfoPersonal);
-		return new ResponseEntity<>(HttpStatus.OK);
+		
+		if(service.findById(idInfoPersonal)!=null) {
+			service.delete(idInfoPersonal);
+			return new ResponseEntity<>(
+					"¡Información personal eliminada!",
+					HttpStatus.OK);
+		}
+		
+		return new ResponseEntity<>(
+				"¡No existe información personal con el ID: "+idInfoPersonal+"!",
+				HttpStatus.OK);
 	}
 }
